Use builtin uglifyjs-webpack-plugin
So far XSnippet Web depends on uglify-webpack-plugin v1.0.1 that depends
on uglify-es v3.1.3 but since the latter is added to dependencies with ^
mark it means we can use any minor version (i.e. any 3.x).

Unfortunately, uglify-es 3.3.0+ (released 2 days ago) broke something
and any minified build currently produced doesn't work and fail with JS
errors in browser console. One way to fix it is to try to pin uglify-es
to latest working version which is v3.2.2. Another is to use builtin
uglify-webpack-plugin which is old (v0.4.2).

This commit implements the latter approach and use builtin version
instead. The rationale is simple: there's no need to use latest version
and we are more than happy with tested bundled version. :)