Do not require propTypes for React components
Each react component can optionally define propTypes which is basically
a schema that's used to validate input props. Despite being optional our
ESLint setup requires those to be defined for each React component. This
is annoying due to the following reasons:

 * I don't believe in type checking in dynamic languages.
 * There's no type schemas for third party libraries.

Taking into account that contracts can be checked in tests and propTypes
are checked only in development version, I suggest to do not enforce
them and make them optional.