Add redirect to newly added snippet
Since we don't have notification about successfully added snippet
it was a best choise to redirect to snippet details after submitting.
To redirect we need to have id of new snippet which is generates on
backend side, thus we need to call redirect after new snippet is
created. We reach this just to pass callback to action.