Add Jest integration along with simple test case
As the project growths, it becomes harder to ensure that each commit
does not introduce regression. One way to have at least some guarantees
is to cover all code base with tests and run them on each PR.

This commit adds Jest (test runner) as well as a simple test to prove
it works and to set an example.