Use immutable.Map() as initial state for store
We already use immutable.js classes in state to be protected from
accidental state mutation. However, the protect is not complete as the
root object is still plain Object() instance. So let's complete
protection by using immutable.js Map as root object for state.

Besides, having root state as immutable.js Map we make it easier to
convert the whole state to plain JS in tests.