Fix ability to copy-paste code from snippet page
In case of CodeMirror has readOnly 'nocursor' option the text cannot
be copied because CodeMirror loose its focus. That basically menas we
can't use this option and we need to use some sort of workaround. One
way to solve this is by using readOnly true option and hidding
the blinking cursor with css rule.