Preserve communication protocol in requests to API
Brand new XSnippet is going to be served using HTTPS protocol, however,
we use hardcoded http:// protocol on API URIs and this is a deal
breaker. Apparently, CORS requests made from https:// to http:// that
redirects to https:// is not going to work.

So we probably will go with both http:// and https:// protocols for API
for now in order to do not break local development environment, and in
order to be production/development compatible let's use protocol-less
schema in API URIs (e.g. //api.xsnippet.org) which means preserve schema
from the current page.