Replace CodeMirror with AceEditor
AceEditor is much more functional that CodeMirror as well as there are
bunch of extensions/styles for the AceEditor in the Internet. For
instance it provides an API to annotate lines or highlight them, which
CodeMirror lacks of.

This commit replaces CodeMirror with AceEditor, though some style
enhacenements are required before going live.