Explicitly require textmate theme
Apparently Ace (Brace) editor uses TextMate color scheme by default and
we used to rely on this. However, if no theme is imported explicitly,
Ace will try to dynamically fetch it and will end up with 404 error in
browser's console, since we don't have the file and everything in
bundled in one file.