Set proper links (hrefs) for "raw" snippet button
Raw snippet functionality is implemented not at SPA side (xsnippet-web)
but at xsnippet-web-backend side. Hence, it may be deployed somewhere
else and there's no "right" endpoint to it. That basically means we need
a way to pass this URL (or URL template) build time, and this is what is
done in this patch.