Do not pass node_modules to Babel
We used to pass Joi and its dependencies to Babel because they are
distributing as ES6 libraries, not as ES5. While browsers are usually
good enough to eat this, UglifyJS fails to process non ES5 blocks.

Fortunately, we've upgraded to webpack 4 that comes with new version of
uglifyjs-webpack-plugin that depends on uglify-es module rather than on
uglify-js. The former supports non ES5 blocks so from now on any non ES5
code processed by uglifyjs-webpack-plugin will be transpiled to ES5.