Probe runtime config before passing control to app
In order to use released XSnippet Web in various environments, we need
to support so called runtime configuration. Basically, it's nothing more
but an attempt to download some JSON and merge its values with default
settings. Runtime configuration is quite handy and could be useful, for
example, when you want to use another (non default) API server.