Split syntaxes out of vendors bundle
AceEditor's modes (aka syntaxes) are pretty heavy, and since they are
not essential, we better download them asynchronously when the app is
loaded. In order to do that we need to split syntaxes out of vendors
bundle into separate one and to initiate its downloading when the app
is loaded.